created schema for user searching, and change logic for user manager

+++ app/handlers/account.py
@@ -5,12 +5,12 @@ from sqlalchemy import update
 from sqlalchemy.exc import NoResultFound
 
 from models.account import Account
-from schemas.account import UserRegistrationSchema
+from schemas.account import UserRegistrationIn, UserSearch
 
 
 class AccountManager:
     @classmethod
-    def create_user(cls, session: Session, data: UserRegistrationSchema):
+    def create_user(cls, session: Session, data: UserRegistrationIn):
         new_account = Account(
             email=data.email,
             full_name=data.full_name,
@@ -21,25 +21,24 @@ class AccountManager:
         return new_account
 
     @classmethod
-    def get_user(cls, session: Session, search_data: tuple):
-        # search data is a tuple with (field_name, value_to_search) data
-        #! Need to change tuple to smth more appropriate
-        if search_data[0] == 'id':
-            return session.get(Account, search_data[1])
-        elif search_data[0] in ['email', 'username']:
-            field = getattr(Account, search_data[0])
-            return session.query(Account).filter(field == search_data[1]).one_or_none()
+    def get_user(cls, session: Session, user: UserSearch):
+        # Works only with id, username and email
+        if user.field == 'id':
+            return session.get(Account, user.value)
+        elif user.field in ['email', 'username']:
+            field = getattr(Account, user.field)
+            return session.query(Account).filter(field == user.value).one_or_none()
         raise AttributeError('Wrong field name')
 
     @classmethod
-    def update_user(cls, session: Session, search_data: tuple, data: dict[str, Any]):
-        # Here I also use search data but I need to switch it to smth better
-        #? but I need to keep in head only one thing, here I use only id value
-        if search_data[0] != "id":
+    def update_user(cls, session: Session, user: UserSearch, data: dict[str, Any]):
+        # Work only for id field as a search field and later I need add validation
+        # for fields
+        if user.field != "id":
             raise ValueError("Update allowed only by id")
         stmt = (
             update(Account)
-            .where(Account.id == search_data[1])
+            .where(Account.id == user.value)
             .values(**data)
             .returning(Account)
         )
@@ -49,8 +48,10 @@ class AccountManager:
         return result
 
     @classmethod
-    def delete_user(cls, session: Session, search_data: tuple):
-        to_delete_account = AccountManager.get_user(session, search_data)
+    def delete_user(cls, session: Session, user: UserSearch):
+        if user.field != "id":
+            raise ValueError("Update allowed only by id")
+        to_delete_account = AccountManager.get_user(session, user)
         if not to_delete_account:
             raise NoResultFound('Account doesn\'t exists')
         session.delete(to_delete_account)
